requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 24.9.2
altgraph=0.17.4=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.7.2=haa95532_0
comtypes=1.4.8=pypi_0
customtkinter=5.2.2=pypi_0
darkdetect=0.8.0=pypi_0
expat=2.6.3=h5da7b33_0
libffi=3.4.4=hd77b12b_1
openssl=3.0.15=h827c3e9_0
packaging=24.1=pypi_0
pefile=2024.8.26=pypi_0
pillow=10.4.0=pypi_0
pip=24.2=py312haa95532_0
psutil=6.0.0=pypi_0
pyinstaller=6.10.0=pypi_0
pyinstaller-hooks-contrib=2024.8=pypi_0
pynput=1.7.7=pypi_0
pystray=0.19.5=pypi_0
python=3.12.4=h14ffc60_1
pywin32-ctypes=0.2.3=pypi_0
setuptools=72.1.0=py312haa95532_0
six=1.16.0=pypi_0
sqlite=3.45.3=h2bbff1b_0
tk=8.6.14=h0416ee5_0
tzdata=2024a=h04d1e81_0
vc=14.40=h2eaa2aa_0
vs2015_runtime=14.40.33807=h98bb1dd_0
wheel=0.43.0=py312haa95532_0
xz=5.4.6=h8cc25b3_1
zlib=1.2.13=h8cc25b3_1
